+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   - Inlining the last binding of a let removes the let. #210
   - Allow import java classes via code-actions `Add import 'x.y.Z'`.
   - Make `add-import` command/code-action smarter, checking if there is already a package import to just include the classname instead of full import.
+  - Add new refactoring `cycle-keyword-auto-resolve`, e.g. `::bar` -> `:foo/bar` and vice-versa. #1128
 
 - API/CLI
   - Allow specify analysis type for `dump` command, default to analyze project only analysis. #1383